Issue #11163: Enforce file size for ConstantName input files
